+++ packages/flutter_tts/example/integration_test/flutter_tts_test.dart
@@ -1,5 +1,3 @@
-// @dart = 2.9
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_tts/flutter_tts.dart';
 import 'package:integration_test/integration_test.dart';
@@ -9,7 +7,7 @@ void main() {
 
   testWidgets('Can speak', (WidgetTester tester) async {
     final flutterTts = FlutterTts();
-    var result = await flutterTts.speak("Hello World");
+    var result = await flutterTts.speak('Hello, world!');
     expect(result, 1);
   });
 }